pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-nrel"
description = "A schema package plugin for labs at nrel and Uni Boulder."
version = "1.0.0"
readme = "README.rst"
requires-python = ">=3.10"
authors = [
    { name = "Michael Götte", email = "michael.goette@helmholtz-berlin.de" },
    { name = "Paolo Graniero", email = "paolo.graniero.goette@helmholtz-berlin.de" },

]
maintainers = [
    { name = "Paolo Graniero", email = "paolo.graniero.goette@helmholtz-berlin.de" },
    { name = "Michael Götte", email = "michael.goette@helmholtz-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.10",
    'nomad-baseclasses @ git+https://github.com/nomad-hzb/nomad-baseclasses.git',
    'nomad-measurements']
[project.urls]
Repository = "https://github.com/nomad-hzb/nomad-nrel"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.uv]
extra-index-url = [
  "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint 
    "PL",
]

pylint.max-branches = 20
pylint.max-statements = 65
ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLR2004"
]


fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
nrel_package = "nomad_nrel.schema_packages:nrel_package"
nrel_jv_parser = "nomad_nrel.parsers:nrel_jv_parser"
nrel_experiment_parser = "nomad_nrel.parsers:nrel_experiment_parser"
